Revert "add dump cs meta"
