+++ Makefile
@@ -139,8 +139,8 @@ mark: $(MARK_OBJ) $(LIBS)
 logdb_dump: src/nameserver/logdb.o src/utils/logdb_dump.o
 	$(CXX) src/nameserver/logdb.o src/utils/logdb_dump.o $(OBJS) -o $@ $(LDFLAGS)
 
-bfs_dump: src/utils/bfs_dump.o
-	$(CXX) src/utils/bfs_dump.o $(OBJS) -o $@ $(LDFLAGS)
+ns_dump: src/utils/ns_dump.o
+	$(CXX) src/utils/ns_dump.o $(OBJS) -o $@ $(LDFLAGS)
 
 bfs_mount: $(FUSE_OBJ) $(LIBS)
 	$(CXX) $(FUSE_OBJ) $(LIBS) -o $@ -L$(FUSE_PATH)/lib -Wl,-static -lfuse -Wl,-call_shared -ldl $(LDFLAGS)
